requirements.txt
flask
psycopg2-binary
flask-sqlalchemy
sqlalchemy
sqlalchemy_utils
python-dotenv
flask_cors